Refactor BottleSong class and improve verse generation
This commit refactors BottleSong class by renaming `top` and `last` to `firstHalf` and `lastHalf`. Also, the first half of the verse is now repeated within the 'firstHalf' function itself instead of in the 'verse' function. This refactoring simplifies the 'verse' function and makes the code more intuitive and easier to follow.