Add domino chain solver and generations for Captain's Log
This commit includes the the Java solution for the creation of random generators for data that commonly appear in Star Trek's Captain's Log. The solution includes different scenarios for planets encountered, starships met and stardates. Also added is the implementation for a domino chain solver game. Updated relevant tests, readmes and gradle files.