Add GottaSnatchEmAll and Dominoes exercises to Java track
This commit introduces two new exercises to the Java track. For the GottaSnatchEmAll exercise, instructions, methods, tests and helper classes have been added to manage a Blorkemon cards collection using the Set interface. The Dominoes exercise involves the creation of Domino and Dominoes classes with appropriate methods and tests for domino chain formation.