Add Annalyn's Infiltration quest logic
This commit adds the Annalyn's Infiltration quest logic into the RPG game. The logic includes implementation of actions such as fast attacks, spying, signalling the prisoner and freeing the prisoner, based on the awake/asleep states of characters in the game and the presence of Annalyn's pet dog. As part of the commit, tests for the quest logic have also been added.