Add CircularBuffer class with associated tests
This commit introduces a new CircularBuffer class, an associated BufferIOException class and a set of unit tests for the CircularBuffer. A circular buffer is a data structure that uses a single, fixed-size buffer as if it were connected end-to-end. The tests are designed to verify the basic functions of the buffer like write, read, overwrite and clear. A build.gradle file has also been added to manage dependencies.