Refactor Ledger formatting logic and introduce ChangeFormatter
Refactored the code for formatting financial entries in the Ledger class to use java.text.DecimalFormat, resulting in a more straightforward and readable code. Additionally, the logic for converting amounts has been moved to a new class called ChangeFormatter, with property files containing formatting patterns introduced in the process. This increases code modularity, enhances readability, and allows for easier maintenance. Changes in the LargestSeriesProductCalculator class were minor with some condensing of expressions.